Wiwynn/DeltaLake: Updated u-root to release v0.9.0 from github
This change brings the u-root v0.9.0 to the DeltaLake coreboot
building process. The master branch on u-root is renamed to
main branch and the hash tag is updated to the v0.9.0
commit on u-root main branch.
This change also required to update go package from 16.6 to 17.1
as that was needed to compile u-root.
The appropriate shasums were also updated for verification.

Tested: Tested booting on DeltaLake with these changes.

Signed-off-by: Eddie Vas <aeddiesharma@fb.com>